Exclude withChangelog from docs
